list sexpr syntax
wraps a List value with an Apply type to change evaluation behavior (in
the future anyway, when that's implemented)

when evaluated, it will Call its first pair value, which should be an
Operative or Applicative.

undecided yet on how () should evaluate - for now I documented it as an
error, but it could be an alias for [] too.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>