Pair -> InertPair, Apply -> Pair, add pair syntax
attaching the evaluation behavior to a discrete type only constructible
with `()` syntax causes friction when dynamically constructing
expressions to be evaluated.

by inverting the types and making `[]` be a special case rather than
`()`, we adopt the conventional evaluation semantics for the Pair type,
leaving `[]` to be a new type, InertPair, which evaluates its contents
and returns them as a Pair.

along the way, split out some types into separate files so they match
their test files. (applicative, pair, empty)
Signed-off-by: Alex Suraci <suraci.alex@gmail.com>