add Assoc, Object, and Keyword types
a Keyword is a constant value similar to a String. the key difference
(heh) is that a Keyword :foo-bar has an internal value of "foo_bar" when
translated into JSON.

an Object is a mapping from Keywords to Values. i decided to stick with
Object as the name instead of Map, because the keys must be Keywords
(Clojure accepts any value), which maps more cleanly to JSON.

an Assoc is a syntactic form to construct an Object, similar to Cons for
.Pair. structurally it is a sequence of Pairs whose first slot must
evaluate to a Keyword.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>