add pipes (sources and sinks), use them for stdi/o
this introduces the concept of 'pipes', which represent values passing
between parties as JSON streams - like UNIX pipes, but for structured
data, not just text.

a source is the read end of a pipe - something you can read values from
(or close)

a sink is the write end of a pipe - something you can emit values to,
sending them to all parties downstream (i.e. it may be a fan-out)

values emitted to the stdout sink will be written as a JSON stream on
stdout

values from the stdin source will be read from stdin as a JSON stream

(these will be more useful later on once running commands is introduced)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>