fix backtraces
continuations are now called with (Value, error) instead of just
(Value). this way Annotated can intercept the error.

i kind of expected this to undo some of the performance gains and maybe
cause a growing heap again, but it turned out fine; no observable leak
occurs with the stdio example.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>