revert this: distinguish ReadyCont from Value
overloading the Value return type and checking for a ReadyCont felt
awkward, and I'm not planning to expose continuations to the language,
so I aimed here to split it out as a distinct return value.

but then it led to a growing heap! Annotated kept growing. removing the
&Traced{} allowed it to work without growing the heap, but then there'd
be no backtrace support.

seems I was right to be surprised that the earlier implementation
worked; it's pretty brittle! I would need to do a lot more investigation
to clearly explain why, so I'm just going to revert this for now.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>