add path types
paths are first-class values used to represent paths on the local
filesystem.

paths are distinct types from strings and are not convertible to
strings. (at least, not right now, and it's an interesting constraint to
leave in place.)

future plans: bass scripts will be run from a working directory in which
to write arbitrary data, ignoring the user's initial working directory.
only paths emitted to `*stdout*` will be kept; the rest may be removed.
this pattern allows scripts to create it as a temporary directory,
creating intermediate files that can be ultimately thrown away, without
having to deal with /tmp.

paths may be constructed with literal syntax:

* ./ is a DirectoryPath referring to the working directory
* ./foo is a FilePath referencing the 'foo' file in the working
  directory
* .foo is a CommandPath representing an executable in the runtime
  environment's $PATH
* foo/bar extends the path bound as `foo`, referencing a file 'bar'
  within the directory

notably, absolute paths may not be constructed: an explicit error is
returned. bass are encapsulated within their working directory by
design. absolute paths may only be passed in by a user on the
commandline. (which is not implemented yet.)

i plan to also forbid use of ../ in path syntax, but it's not
implemented yet.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>