Symbol.Decode does not support string
this is a relic, and I'm not sure it was ever needed. all tests stay
passing, so let's keep this conservative.

impact of this: previously, a symbol would be accepted by any builtin
fns that take string arguments. a) there might just not be any, and b)
that's a pretty challenging thing to do accidentally.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>