annotated errors include commentary
this is an experiment with the overall goal being to allow inline
comments to become part of the error feedback loop. drop a comment on a
flaky workload and it'll show up when it fails, which may help yourself
in the future or someone else on the team when troubleshooting the
failure.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>