Path is filepath, Resolve preserves absolute paths
previously, paths used the platform-neutral `path` package and only
returned a platform-specific path when .Resolve is called.

when accepting paths from the CLI, I want it to pass them as fully
resolved absolute paths so that they don't become incorrect when the
script runs from a different directory (in the future, when that's
implemented).

to do this, Resolve should check IsAbs, but path.IsAbs returns false for
a forward-slash'd Windows path (C:/foo/bar). so let's just change Path
to use platform-specific path separator instead, so we can do a
filepath.IsAbs check instead of path.IsAbs.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>