fix tight loops overwhelming the heap with traces
apparently my suspicions around my backtrace fix were justified! it
worked in the stdio case, but failed quickly in a tight loop:

  (defn loop []
    (loop))

  (loop)

this commit replaces the previous tracing method with a more efficient
approach based on keeping a window of the last 100 calls in a trace kept
in context.Context.

that number is completely made up and should probably be tweaked in the
future. haven't given it any thought. it should probably be made larger
to account for deep call stacks which unwind and erase large parts of
the trace.

side note: i played god and rebased so that tail_recursion_test.go was
added along with the original commit that introduced tail recursion, to
aid in bisecting. i'm `push -f`ing with reckless abandon because I like
that it makes it easy to compare test results, and it's accurate in
breaking the build (catching a regression) on all the commits that fail
in between.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>