add json encoding/decoding for supported types
notably this establishes a convention for Decoding Object values into
other structural types (i.e. paths).

linguistically, this would mean any builtin function that takes a
CommandPath, DirectoryPath, or FilePath can also take an object in the
following forms:

  {:command string}
  {:dir string}
  {:file string}

meaning .foo, ./foo, and ./foo/ are sort of syntactic sugar, with
applicative semantics for constructing workloads, but when being sent
over the wire they are isomorphic to the above structures.

admittedly this is a little loosey-goosey, but I think it's a reasonable
compromise for being able to pass them over the wire as JSON; it's
unlikely for these object forms to have any other meaning (given they
only have a single field).

this convention may be used in the future for more complex types (i.e.
workload paths).

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>