support reading absolute paths
this was an opinionated stance aligning for the previous goal of running
things on the host machine and touching the local filesystem

i'm pivoting away from that and back to the original plan of only
executing reproducible workloads via runtimes and never interacting with
the host machine.

with the original risk gone, this should actually be allowed in support
of referencing absolute paths within the container that the workload
runs in, which is needed to support things like the Concourse resource
protocol (/opt/resource/{in,out,check}).

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>