Revert "Path is filepath, Resolve preserves absolute paths"
This reverts commit f2b7a64bc99081114b01b45b0335fd844361f9f5.

going to keep paths platform-neutral so that they can be interpreted by
the runtime instead; bass running on Windows might submit workloads to a
Linux runtime, so it wouldn't make sense for it to receive Windows
formatted paths.

the original use case justifying this change could maybe be done with
the addition of a distinct HostPath type instead.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>