switch to exclusively json: tags
previously, for an object that exists both in bass and over the wire
(i.e. to a runtime), you had to specify three tags:

  bass:"foo" optional:"true" json:"foo,omitempty"

this changes Decode and ValueOf to just interpret the json: tag instead.
this seems less error-prone and a lot easier to maintain, and we're
already pretty married to JSON, so so this seems like a pragmatic
choice even though it is a little tightly coupled.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>