remove ID from workload
the runtime CLI will have a few commands for interacting with workloads,
so this will be passed as an argument instead of requiring a full
workload to be passed on stdin.

this field also really shouldn't be configurable by the user; it breaks
the abstraction and could cause problems. so in the future the ID will
be generated, probably as a digest of the workload.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>