implement runtime system for running workloads
this is a bit of a big bang. it took a few iterations to get here so it
was pretty messy and bloated and confusing partway through.

there's a lot to cover but it's probably better to spend that effort on
the docs.

tl;dr: runtimes run workloads and cache their results. currently a
docker runtime is implemented. additional runtimes may be implemented in
the future.

side note: at one point I had runtimes implemented as standalone CLIs. i
decided to go with having them all built in since it's a lot simpler
that way, and i think it'll be an overall better user experience.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>