Workload.Name -> Workload.SHA1, SHA256, etc.
Name was a bit of a misnomer - these workloads are meant to be
content-locatable, independent of any name attached to it.

adding various methods for hashing allows runtimes to use different
hashes for different contexts (i.e. string length constraints) while
avoiding blessing any particular hash as the naming scheme.

caveat: sha1 is used in WorkloadPath.String, so it's somewhat blessed.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>