remove platform argument from resource
all resources are linux, and that might forever be the case; the
interface is fundamentally coupled to Linux (/opt/resource/*). resources
ought to be primarily just doing network I/O so there isn't much point
in implementing them on other platforms. generally building should be
done with tasks (or just regular old workloads) instead.

prototypes might be multi-platform in the future, but that would
be implemented by a different interface.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>