switch to buildkit's progress ui kit
pretty much perfect for this use case:

* tracks timing info so you can clearly see the slow parts
* makes it possible (and intuitive) to show logs for multiple workloads
  in parallel
* supports breaking down a vertex into multiple sub-tasks with their own
  timing info
* looks cool

hats off to all who worked on it, console UIs like this are a pain in
the butt to implement.

i hope to extract a package for it, since I'd like to tailor its UI a
bit more for bass.

there were definitely some tricky parts to this - but I think most of it
just came down to race conditions (thank god for --race) and misuse of
the internal API (had to reverse-engineer it since it's understandably
not documented).