fix runtime errors not failing the vertex
tricky named return value + defer interaction led to err being clobbered
with the err result of removing the container, which pretty much always
succeeds.

instead we just record the remove error on the vertex. caveat: the error
will be clobbered by the original err if there is one, but that doesn't
seem so bad. a multierror would be nice, but it  doesn't seem worth the
hassle.