tweak resource constructor to use scope literal
using (get-current-scope) like this was a bit of a footgun; not only
would it export the definitions in the current scope, it would
technically export all definitions in the inherited outer scope.

this definitely doesn't preserve the doc strings, but it probably didn't
work very well before either. it'd be nice if there was a shorthand for
extracting a subset of symbols from a scope to an isolated scope, along
with their docs.