runtime: pass relative paths, accounting for cwd
initially, all workload path values passed to commands were relative to
the initial working directory. this would break however when passed to a
command with a different working directory configured (perhaps another
workload path), because they would be interpreted relative to the
command's configured working directory.

i "fixed" that by making all paths absolute, which I think is the
simplest solution, but that caused breakage with the Github Release
resource, which does a blind filepath.Join(sourceDir, namePath) causing
the path to become a bogus relative path (/tmp/... => tmp/...).

technically the resource should be fixed to respect absolute paths, but
i'd rather try to just be compatible out of the box; who knows where
else this mistake is made.

this commit goes back to relative paths, but carefully makes them
relative to the configured working directory instead.

that is, running a command with dir set to some-wlp/some-dir/, and
another workload path other-wlp/some-file passed to argv or stdin, it
will resolve to ../../other-wlp/some-file instead of
./other-wlp/some-file.

note that this transformation only applies to workload paths, which are
just expected to exist in some location. file or directory paths are
taken literally, just as before.