concourse: limit put cache duration to 1m
configurable with :put-cache-duration

sometimes you need to re-run a put because external state changed or was
reset. put steps should be idempotent, so it should always be safe to
re-run it anyway.

a duration of 1 minute was chosen semi-arbitrarily; it could be nice to
have a brief window of time just to de-dupe concurrent runs of the same
workload, in case the runtime implements that.