replace doc/comment system with ^metadata
initial motivation: annotating stdlib definitions to inform editor
config and docs generation.

for example, annotating an operative with {:indent true} to indicate
that it should have `set lispwords=...` style indentation:

  ; construct an operative and bind it to a symbol
  ^:indent
  (defn defop ...)

* Annotated has been split/renamed into Annotate and Annotated
  * Annotate evaluates into Annotated, a pairing of a value with a
    metadata scope
  * Comment is set as a :doc metadata field
  * Range is set as :file, :line, and :column fields - but only for
    annotated bindings
* new reader macro for metadata: ^meta form
  * evaluates form followed by meta, returns an Annotated value
  * meta must be a {bind}, :keyword, symbol, or "string"
    * same sugar as Clojure
    * {:a 1} is a metadata scope
    * :foo is shorthand for {:foo true}
    * "foo" is shorthand for {:tag "foo"}
    * foo is shorthand for {:tag foo}

this system is inspired by Clojure but the way it works is pretty
different. any value can be annotated, not just symbols and collections.
metadata is passed by value instead - a scope paired with the value it
annotates.

metadata can be paired with a value using (with-meta val meta), or be
set literally using the ^meta reader macro.