remove env commentary, use *doc* instead
this includes a bunch of other churn wherein i went back and forth on
having vito/is use go-cmp/cmp for comparison, ultimately going back to
the Equal helper I had before this whole switchover.

pardon the mess.

the actual change here is to get rid of Commentary from scopes and
simplify down to just setting metadata on bindings. instead of recording
commentary into the scope, a scope should define a *docs* binding. the
binding's own docs (comments) represent scope-level docs, where you may
put an intro/broad description. the binding's value is ignored for now -
but in the future it could be a structure categorizing the bindings.