remove github actions stuff
there was too much of an impedance mismatch; github actions are tightly
coupled to github's bespoke environments and are extremely repo centric.
there's a lot to re-implement.

the original motivation was to be able to re-use actions people have
already written, but IMO this is less of a problem now that images are
so easy to build and use within bass.