github: init submodules
this is a downside of the previous change: now that we're feeding local
files directly into bass, the results aren't guaranteed to be the same
between CI and dev. they passed locally because i had already
initialized the submodules, but failed repeatedly in CI. irritating to
debug!

suggestion: scripts in hack/ use local files, scripts ci/ control all
inputs.