bass runtime: (run) doesn't assume .bass extension
it's tempting to try both based on which one exists, but that becomes
difficult with thunk-provided libraries, and makes things a little less
predictable.

instead, bass scripts intended to be executed should omit the .bass
extension, and scripts intended to be loaded must include the .bass
extension.

this follows patterns in how i've used it so far, so i'm ok with this,
but we'll see if people find it too confusing.