all paths are now combiners
'extend' syntax is now shorthand for a chain of function calls, each
receiving a file or path argument, rather than a chain of path extend
calls.

in other words, extend syntax:

    foo/bar/baz/buzz

is now shorthand for this:

    (((foo ./bar/) ./baz/) /buzz)

and now, foo can be any combiner.

this was the truth table before:

    f = file, d = dir, c = command, tf = thunk file, td = thunk dir
    0 = supported, 1 = unsupported

    | supports | c  f  d tf td hf hd |
    | Extend   | 0  0  1  0  1  0  1 |
    | Call     | 1  1  0  1  0  1  0 |

now all paths implement Call, and they just implement different
behavior.

    f = file, d = dir, c = command, tf = thunk file, td = thunk dir
    t = create thunk, s = subpath

    | supports | c  f  d tf td hf hd |
    | Call     | t  t  s  t  s  t  s |

granted, it's a lateral move, but the advantange of just unifying on
Combiner is now anyone can implement their own path types:

    (defn github [user]
      (fn [repo]
        (let [uri (str "https://github.com/" (name user) "/" (name repo))]
          (fn [route]
            (case route
              ./sha/
              (fn [sha] (checkout uri sha))

              ./ref/
              (fn [ref]
                (checkout uri (ls-remote uri (name ref)))))))))

    ; resolves to a thunk path of the checked out repo
    github/vito/bass/ref/master