add (resolve) for resolving an image ref
adds a new runtime interface, Resolve(), which takes an ImageRef and
populates its Digest field.

the plan is to make thunks more hermetic by resolving image refs to an
exact digest before they're even used by a thunk.

the original tag is kept alongside the digest so that the user can
choose to resolve it again if the digest disappears for some reason
(e.g. pushing over 'latest' and registry GCs old digest).