respect buildkitd worker platform, cache resolving
previously, image resolution would happen on the client-side when
building the LLB definition; now it uses the gateway for image
resolution, and specifies the platform of the buildkitd workers.

in addition, Resolve now goes through the cached resolver which should
speed up image resolution. at some point it might be better to stash
this in a .lock file of some sort.

part of #35