add support for secrets
this is a simple first step: it adds a secret type, and makes it safe to
pass secrets to thunks, but it doesn't address how to safely pass these
secret values to thunks in the first place. that'll come next.

secrets values can now be safely (knock on wood) passed to commands in
env, stdin, argv (but please don't[^1]), and via secret mounts.

this works by leveraging the existing shim - the command is passed to it
via a JSON file mounted in, which the shim burns after reading. this
prevents them from being captured in the image layers.

first-class secret values can be constructed like so:

    (make-secret :bruce-banner "im always angry")

a secret may be passed around as if it were its contained string value.
thunks may embed secrets, whose value will be used by the runtime, but
omitted when encoded to JSON - leaving only the name. this makes it safe
to publish thunk paths even if secrets were involved in its creation.

secrets passed to commands via argv will be masked in the progress
headers like so:

    sh -c "echo $0" <secret: secret-name (6 bytes)>

note that this is not a generic output mask. it works by rendering the
thunk's values instead of the command's values, and secrets hide their
value. it won't help if your `set -x` prints it.

[^1]: secrets probably shouldn't be passed in argv, since it can end up
in fishy places like the shell history.

part of #7