add virtualized env vars
env vars are now available in *env* - with a catch: they are only
accessible to the entrypoint script. the 'real' system env vars are
cleared after being converted to a bass scope, so env vars must be
explicitly passed around from that point onward.

so to safely pass secrets into a bass script, just set an env var read
by the entrypoint. it will be impossible for other things to access it
without it being explicitly passed to them (i mean, unless there's some
other exploit).

accordingly, env vars can now be set in bass thunks, which is the only
way to pass them. if not set, *env* will be an empty scope.

one nice outcome of this is unbound env vars will result in a decently
obvious error without having to do anything special.