add (memo) and (unmemo)
this seems to be enough for an MVP of version pinning. i have no idea if
this will be the long-term plan.

(memo) memoizes a function, storing categorized input/outputs in the
given path. if the path is a dir, it will be traversed upward until a
bass.lock is found. if the path is a file, it'll be used directly and
created if it doesn't exist.

(unmemo) lets you forget. it takes the memos path, a category, and an
input, and removes the input's memoized output. filtering or clearing
categories can be added in the future.

with these two added...:

* `linux' now memoizes by locating the nearest bass.lock to the caller's
  *dir*. refs.db is no more, and buildkit runtime no longer has to
  implement caching!
* `git:github' does the same - now you can efficiently point to tags, or
  even dynamic branches, and have stable dependencies.

using the caller's *dir* is kind of sneaky, but it really helps with
bootstrapping, since these are the two first things any bass script is
likely to do, and I really don't want to have to explicitly initialize
them with a `bass.lock' all the time.

TODO:

* support reading bass.lock from within a thunk path (e.g. a fetched
  bass lib)

bonus:

* git paths now support taking strings in addition to paths, so checkout
  can be implemented like (git:github/vito/bass sha). pretty slick!