bass: pass bass binary to tests thunk, optimize
previously the tests thunk would just run `make install` to build a bass
binary to test against[^1], but the build thunk ignored the Makefile and
just ran `go build`.

now the build thunk is an input to the tests thunk, so we're testing the
same thing we're shipping.

`make` is no longer used for building `bass`, though the Makefile lives
on for bootstrapping purposes.

also, some optimizations:

* extract 'go mod download' setup to a thunk that can be shared by shim
  build, bass build, and tests
* pass only the shim source to the shim builder so it can be cached more
  aggressively

[^1]: primarily the lsp tests, since the rest are in-process.