replace lock 'category' with thunk + binding
bass.lock now stores not only the results for a function, but everything
needed to re-run it if desired.

instead of a category, each input/output is associated to a Bass thunk
to (load) and the binding of the function to call.

this should be much more airtight.

along the way, replace (memo) with a native bass implementation in terms
of two more low-level bindings (which I might move to the internal
scope).