optimize shim memory usage
heap profiling showed a lot of memory being used here; it's better to
just load them into memory once and pass the content in.

the actual impact of this is pretty low; I was actually looking for
something else, and this was a red herring, but it's worth fixing up
anyway.