normalize timestamps to 1985-10-26T08:15:00Z
file mtimes are kryptonite for reproducible builds. this change extends
the shim to set the atime/mtime for each file in the thunk's output
mount to the following time:

  October 26, 1985 at 1:15 AM PST

or as a Unix timestamp, 499162500.

as a result, bass's self-build is now reproducible!

  $ ./ci/build 1ff7a699621d9bdad718ee096984ed2c952bdaa4 linux amd64 | bass -e | sha256sum
  ad87286daf4c614f1948a837f88b5cd1674a7a71d7851beecc1c75c130fc1064
  $ bass --prune
  $ ./ci/build 1ff7a699621d9bdad718ee096984ed2c952bdaa4 linux amd64 | bass -e | sha256sum
  ad87286daf4c614f1948a837f88b5cd1674a7a71d7851beecc1c75c130fc1064

this arbitrary timestamp is the same one used by npm. notably it isn't
the Unix epoch since some tools (i.e. zip) don't behave well with that.