much nicer thunk errors, error printing in general
on error, finish the progress UI and print the error, instead of
printing the error into the progress UI. this might result in weird UI
when piping bass scripts to each other, but it feels a whole lot
cleaner.

also, display a summary of the progress emitted by a thunk.