add (across) and (for)
not 1000% sure about the names for these.

(across) takes multiple streams and returns a single stream that returns
values read from each stream. can be used for independent triggers in a
pipeline. it doesn't skip values on individual streams, but it also
doesn't produce every permutation.

(for) is syntax sugar for iterating over multiple streams.

  (for [even (list->source [0 2 4])
        odd (list->source [1 3 5])]
    (logf "for: %d %d" even odd))

building blocks for pipelines!