unify ci/ and hack/ under bass/
previously, ci/ was for things "run by ci" and hack/ was for things "run
by devs". ci/ would typically prefer to start with a SHA and fetch a
reproducible input, while hack/ would always run with host paths.

with the last slew of changes it's pretty easy to unify them so they
take a `src` value on stdin, which can be easily passed with -i src=./
during dev.