replace server mode with (http:serve)
the cgi-bin-esque 'serve a directory' model was cute but quickly felt
clunky. I wanted to be able to refactor parts of handlers out into
libraries, but putting them near the handler meant serving them as
endpoints too. it was also slow, since it would parse the file each
time.

now server mode has been replaced: the -s flag is no more, and instead a
http stdlib has been added, currently defining only (listen), which
takes an addr and a handler:

  (http:listen "0.0.0.0:8080" (fn [req res] (res {:ok true})))

there are a few advantages to this:

* requests just call the handler instead of loading a module each time
  * no more parsing overhead
* handlers can return early by calling (res) and carrying on
  * the value will be sent as JSON
  * before you had to (start) a thunk, so you had to implement async
    behavior in a separate file
* much easier to implement handlers which have an empty response
  * just don't call (res)