support manually running checks
relying on a server is annoying since it has to be running somewhere and
you have to keep it up-to-date. it's also a risk vector.

so now when I'm too lazy to set up the server, or if it was down and an
event was missed, I can just run the checks myself.

theoretically this could behave sort of like serverless CI: contributors
would run their own checks, which would link to the thunk that they ran
in some content-addressible server. the server could show the thunk's
output, and the the upstream maintainer can then verify the thunk by
inspecting it or running it themselves.