remove 'server mode' experimentation
I was able to successfully get a GitHub app running and handling
webhooks, which was pretty cool, but I don't think this really belongs
directly in Bass; it's too high-fidelity.

I would rather build a GitHub app that integrates in more compelling
ways than a simple webhook handler. For example, I want to be able to
record and save output/progress to a database and render it in a web UI.
It would be simpler to just build this app in a proper language and have
it call out to Bass where it needs to.