stop relying on /../ hack for host paths
this is an antipattern as it goes against the usual mantra of host paths
which is to be strictly limited to their context dir.

it was always hokey, and this arrangement makes it unnecessary, so I'm
keen to remove it for consistency's sake.

the fewer files required in the toplevel of the repo, the better.