Bass runtime is no longer a runtime
this was never a great fit; half of interface was not relevant to it,
and half of it was not relevant to real runtimes

I'd like to scope the Runtime interface down to "real" runtimes now and
have it just use protobuf types directly. at that point it really makes
no sense for the Bass runtime to be a runtime.