get rid of automatic (un)hyphenation
this is just too surprising. it was originally motivated by having
params sent to Concourse resources feel more native, e.g. :private-key
instead of :private_key. it was a funky idea I went forward with to see
how it went, and to be honest I don't think it went that well.

there were a few cases where I *didn't* want it to happen, and the
translation was destructive in those cases. for example, using the
base64-encoded sha256 of a thunk as a key in bass.lock would become
corrupted.

it's better to just have this match exactly what the user says. if
there's some ergononomic reason to do this conversion it should be
applied in a much more localized area (e.g. a Concourse lib) rather than
the entire language.