eliminate duck-typed JSON pattern (Descope)
this was always a little bit of a wild west, so it's a little comforting
seeing it go away.

thunks and thunk paths now marshal to JSON using protobuf encoding, but
they - along with other rich path types - will no longer be
automagically translated from their JSON encoding. this always risked
ambiguity, with it being unclear whether a {"file":"./foo"} form should
actually turn into a bass.FilePath or a general *bass.Scope.

now that thunks and memos encode using protobufs, this ambiguity is
gone. emitting a thunk path as JSON to stdout and piping to `bass
--export` still works just fine because --export knows it's expecting a
bass.ThunkPath.