keep sensitive secrets out of protobuf
this is a fundamental requirement, as we don't want secret rotation to
change the 'identity' of a thunk, and we don't want to send secrets over
the wire in the first place.

the next iteration on this may change secret names from simple strings
to some sort of namespacing scheme so that user-provided runtimes can
fetch them from a local secret store.