(start) calls handler with error combiner instead
i already shot my shot with the 'a nu start' joke, but it'd be better
applied here. oh well.

previously (start) would call the function with a bool indicating
success. if the handler raised an error, it would actually be used to
wrap the root error. this was always a little surprising.

meanwhile the handler had no way to access the root cause of the error,
e.g. to log it using cli.WriteError(ctx, err), which I need to do in
Bass Loop.

now the handler is called with a nullable combiner. if null, the thunk
succeeded. if a combiner is present, it can be called to propagate the
error to whoever waits on the thunk run.

to accomplish this a special Error combiner value has been introduced.
it displays like <error: inner error msg> and when called returns the
inner error.