cache paths are now distinct path types
previously, a 'naked' file or dir path would be treated like a cache
when passed to a thunk. this was a bit hastily slapped together and
it was never really obvious behavior. if anything folks would likely do
this trying to mount a host path in, and have it do _something_, but not
what they expect.

now cache paths are a distinct path type constructed like so:

  (cache-dir "foo")

where "foo" is an explicit identifier for the cache dir. rather than
using the path itself as the cache key, the user provides one for the
root. this means you can now pass a subdirectory of a cache dir to a
thunk if you want - previously that would yield an entirely different
cache key.

you can also use a cache path as a thunk command now, because why not.