add thunk ports and addrs
* thunks can now configure named ports
* ThunkAddr is a new Value referencing a port provided by a thunk

ThunkAddr is like ThunkPath but for TCP addresses instead of files. It
embeds a format string to be resolved at runtime. They are constructed
with the (addr) function.

e.g.:

    (addr (with-ports ($ foo) {:http 8080}) :http "http://$host:$port")

When a ThunkAddr is used in a thunk its embedded thunk is started, all
ports are polled until ready, and the value is replaced with the
rendered format string.

    (let [thunk (-> ($ foo) (with-ports {:http 8080}))
          addr (addr thunk :http "http://$host:$port")]
      ($ curl $addr))

(not a working example; images etc. omitted)

Currently buildkit does not support custom networking so $host will
always be localhost, but support may arrive someday.