(with-env) is additive, nulls act as tombstones
justification
-------------

Here's a function I use to run commands with a Go cache + some sane
default env:

    ; sets up $PATH appropriately and caches dependencies + builds
    (defn with-go [thunk]
      (-> thunk
          (with-env {:GOPATH "/go"
                     :GOCACHE "/gocache"
                     :PATH "/go/bin:/share/go/bin:/usr/bin:/bin"})
          (with-mount (cache-dir "bass-loop/gopath") /go/)
          (with-mount (cache-dir "bass-loop/gocache") /gocache/)))

(from Bass Loop)

The problem here is if `thunk` has any env set ahead of time it will be
clobbered by this "thunk middleware" (for lack of a better term).

I think this might become a really common pattern, so having
`(with-env)` replace everything by default feels like a bit of a
footgun.

I think `with-env` should behave more like `with-mount` and add to the
existing env instead. This can be done without mutation by constructing
a new scope with the given env and the previous env (if any) as parents.

alternatives considered
=======================

This could be implemented in terms of `(with-env)` if there was a
`(thunk-env)` for getting a thunk's current env:

    (defn set-env [thunk env]
      (with-env thunk {env (thunk-env thunk)}))

but:

* I haven't needed `(thunk-env)` for anything else yet, and felt a
  little weird about adding `(thunk-cmd)` tbh, but `(use)` and
  `(wrap-cmd)` were good justification
* I don't see why you would ever want to replace the entire env as
  opposed to unsetting individual vars

So I would propose we support `(with-env thunk {:FOO null})` to place a
"tombstone" over an env var. This is probably not supported at the
moment. `runtimes/command.go` could probably just skip `bass.Null`
values rather than trying to set them. (An empty string must be
supported and passed through imo.)