fix silly cache race
this was obviously broken! we would check if the cache already existed
and return it, but then create the cache and start writing to it.

now we'll create a temporary file and then rename it over the cache
instead.

this should fix all the really weird errors hitting bass-loop whenever
hooks kick off concurrently.