shim: reap zombies
rather than forcing folks to configure a dumb-init entrypoint, we'll
just have the shim handle reaping zombies, since it's already there, and
it's pretty handy to not have to ever think about this

sidebar: we're getting into "how i learned to stop worrying and love the
shim" territory. it's a bit of a pain to have to maintain, but it sure
grants a lot of flexibility!