don't ignore caching for services
the intention here was to prevent accidentally caching a service that
exits 0 when interrupted.

in reality this also prevented explicit cache mounts from working;
they'd always just be empty. cache mounts are required for services that
keep state, e.g. Nixery.

we also don't want to introduce any randomness to the LLB since that
would prevent sharing of resources.

it doesn't seem possible to meet all of these constraints at the moment,
so just get rid of the cache prevention as I haven't seen it actually
cause any problems yet. maybe because we always interrupt the service,
and that prevents buildkit from caching even if it exits 0? or maybe
everything is just exiting nonzero properly? (none of this is confirmed,
but it would be nice!)