rework runtime test setup for monolithic progress
previously progress output wasn't being captured and displayed by
failing Dagger tests because for Dagger the progress is emitted at the
pool-level, which was constructed ahead of the progress UI being wired
up.

this simplifies the individual runtime tests since they just need to
pass a single runtime config in, but I'd still prefer to not have a
monolithic progress sink. oh well. maybe that'll improve when we figure
out the progress API in Dagger.