don't require secrets in args to be safe
passing secrets in command args is never a good idea. they show up in
the process tree and get saved into shell history. no point in trying to
pretend this is safe.

but really, Dagger doesn't support it, and I think that's OK, so let's
just relax the test a bit.