support building from Dockerfiles
Currently this is its own special type of image, slotting in alongside
OCI archive images, image refs, and thunks. So you don't 'build' a
Dockerfile into a Thunk, you can just use it as an image.

This will allow for far greater compatibility with the external world.
Bass doesn't replace Dockerfiles; it's far too opinionated. Dockerfiles
will probably exist until the end of time if not just as a
language-neutral format for building images.